requirements.txt
allure-pytest==2.8.7
allure-python-commons==2.8.7
allure-python-commons-test==2.8.7
arrow==0.15.5
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cssselect==1.1.0
ddt==1.2.2
easydict==1.9
Faker==4.0.0
idna==2.9
isort==4.3.21
Js2Py==0.67
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
parse==1.14.0
parse-type==0.5.2
phonenumbers==8.11.2
pluggy==0.13.1
py==1.8.1
pycparser==2.19
PyExecJS==1.5.1
pyjsparser==2.7.1
pylint==2.4.4
PyMySQL==0.9.3
pytest==5.3.5
pytest-pycharm==0.6.0
pytest-xdist
pytest-timeout==1.3.4
python-dateutil==2.8.1
pytz==2019.3
tox==3.14.3
PyYAML==5.3
redis==3.4.1
requests==2.22.0
requests-cache==0.5.2
robotframework==3.1.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.2.0
selenium==3.141.0
six==1.14.0
text-unidecode==1.3
tzlocal==2.0.0
urllib3==1.25.8
wrapt==1.11.2
pychrome==0.2.3
websocket-client==0.56.0
webdriver-manager==2.3.0
testcontainers==2.5
asn1crypto==1.3.0
bcrypt==3.1.7
cffi==1.13.2
cryptography==2.8
paramiko==2.7.1
pynacl==1.3.0
